Move helpers method to the separate package
