+++ test/e2e/utils/service.go
@@ -1,0 +1,45 @@
+/*
+Copyright 2018-2022 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"context"
+
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	clientset "k8s.io/client-go/kubernetes"
+)
+
+// CreateService creates nfd-master Service
+func CreateService(cs clientset.Interface, ns string) (*v1.Service, error) {
+	svc := &v1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "nfd-master-e2e",
+		},
+		Spec: v1.ServiceSpec{
+			Selector: map[string]string{"name": "nfd-master-e2e"},
+			Ports: []v1.ServicePort{
+				{
+					Protocol: v1.ProtocolTCP,
+					Port:     8080,
+				},
+			},
+			Type: v1.ServiceTypeClusterIP,
+		},
+	}
+	return cs.CoreV1().Services(ns).Create(context.TODO(), svc, metav1.CreateOptions{})
+}
